1.3.0 a lot of fixes.

+++ AntiCaptchaApi.Net/Requests/Abstractions/Interfaces/IRecaptchaV3Request.cs
@@ -6,5 +6,5 @@ namespace AntiCaptchaApi.Net.Requests.Abstractions.Interfaces;
 public interface IRecaptchaV3Request : IWebCaptchaRequest<RecaptchaSolution>, IApiDomainArg, IIsEnterpriseArg
 {
     public string PageAction { get; set; }
-    public double MinScore { get; set; }
+    public decimal MinScore { get; set; }
 }
